Refactor package media notification
